+++ src/Braintree/WebhookNotification.cs
@@ -22,11 +22,14 @@ namespace Braintree
         [Description("grantor_updated_granted_payment_method")] GRANTOR_UPDATED_GRANTED_PAYMENT_METHOD,
         [Description("granted_payment_method_revoked")] GRANTED_PAYMENT_METHOD_REVOKED,
         [Description("local_payment_completed")] LOCAL_PAYMENT_COMPLETED,
+        [Description("local_payment_expired")] LOCAL_PAYMENT_EXPIRED,
+        [Description("local_payment_funded")] LOCAL_PAYMENT_FUNDED,
         [Description("local_payment_reversed")] LOCAL_PAYMENT_REVERSED,
         [Description("oauth_access_revoked")] OAUTH_ACCESS_REVOKED,
         [Description("partner_merchant_connected")] PARTNER_MERCHANT_CONNECTED,
         [Description("partner_merchant_declined")] PARTNER_MERCHANT_DECLINED,
         [Description("partner_merchant_disconnected")] PARTNER_MERCHANT_DISCONNECTED,
+        [Description("payment_method_customer_data_updated")] PAYMENT_METHOD_CUSTOMER_DATA_UPDATED,
         [Description("payment_method_revoked_by_customer")] PAYMENT_METHOD_REVOKED_BY_CUSTOMER,
         [Description("recipient_updated_granted_payment_method")] RECIPIENT_UPDATED_GRANTED_PAYMENT_METHOD,
         [Description("subscription_canceled")] SUBSCRIPTION_CANCELED,
@@ -39,6 +42,7 @@ namespace Braintree
         [Description("sub_merchant_account_approved")] SUB_MERCHANT_ACCOUNT_APPROVED,
         [Description("sub_merchant_account_declined")] SUB_MERCHANT_ACCOUNT_DECLINED,
         [Description("transaction_disbursed")] TRANSACTION_DISBURSED,
+        [Description("transaction_reviewed")] TRANSACTION_REVIEWED,
         [Description("transaction_settled")] TRANSACTION_SETTLED,
         [Description("transaction_settlement_declined")] TRANSACTION_SETTLEMENT_DECLINED,
         [Description("unrecognized")] UNRECOGNIZED
@@ -56,6 +60,8 @@ namespace Braintree
         public virtual WebhookKind Kind { get; protected set; }
         public virtual MerchantAccount MerchantAccount { get; protected set; }
         public virtual LocalPaymentCompleted LocalPaymentCompleted { get; protected set; }
+        public virtual LocalPaymentExpired LocalPaymentExpired { get; protected set; }
+        public virtual LocalPaymentFunded LocalPaymentFunded { get; protected set; }
         public virtual LocalPaymentReversed LocalPaymentReversed { get; protected set; }
         public virtual PartnerMerchant PartnerMerchant { get; protected set; }
         public virtual string Message { get; protected set; }
@@ -64,7 +70,9 @@ namespace Braintree
         public virtual Subscription Subscription { get; protected set; }
         public virtual DateTime? Timestamp { get; protected set; }
         public virtual Transaction Transaction { get; protected set; }
+        public virtual TransactionReview TransactionReview { get; protected set; }
         public virtual OAuthAccessRevocation OAuthAccessRevocation { get; protected set; }
+        public virtual PaymentMethodCustomerDataUpdatedMetadata PaymentMethodCustomerDataUpdatedMetadata { get; protected set; }
         
         public WebhookNotification(NodeWrapper node, IBraintreeGateway gateway)
         {
@@ -101,6 +109,11 @@ namespace Braintree
             if (WrapperNode.GetNode("transaction") != null)
             {
                 Transaction = new Transaction(WrapperNode.GetNode("transaction"), gateway);
+            }
+
+            if (WrapperNode.GetNode("transaction-review") != null)
+            {
+                TransactionReview = new TransactionReview(WrapperNode.GetNode("transaction-review"));
             }
 
             if (WrapperNode.GetNode("disbursement") != null)
@@ -158,9 +171,24 @@ namespace Braintree
                 LocalPaymentCompleted = new LocalPaymentCompleted(WrapperNode.GetNode("local-payment"), gateway);
             } 
 
+            if (WrapperNode.GetNode("local-payment-expired") != null)
+            {
+                LocalPaymentExpired = new LocalPaymentExpired(WrapperNode.GetNode("local-payment-expired"), gateway);
+            }
+
+            if (WrapperNode.GetNode("local-payment-funded") != null)
+            {
+                LocalPaymentFunded = new LocalPaymentFunded(WrapperNode.GetNode("local-payment-funded"), gateway);
+            }
+
             if (WrapperNode.GetNode("local-payment-reversed") != null)
             {
                 LocalPaymentReversed = new LocalPaymentReversed(WrapperNode.GetNode("local-payment-reversed"), gateway); 
+            }
+
+            if (WrapperNode.GetNode("payment-method-customer-data-updated-metadata") != null)
+            {
+                PaymentMethodCustomerDataUpdatedMetadata = new PaymentMethodCustomerDataUpdatedMetadata(WrapperNode.GetNode("payment-method-customer-data-updated-metadata"), gateway);
             }
         }