+++ test/Braintree.Tests.Integration/TransactionIntegrationTest.cs
@@ -2002,6 +2002,50 @@ namespace Braintree.Tests.Integration
         }
 
         [Test]
+        public void Sale_WithExchangeRateQuoteId()
+        {
+            var request = new TransactionRequest
+            {
+                Amount = SandboxValues.TransactionAmount.AUTHORIZE,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = SandboxValues.CreditCardNumber.VISA,
+                    ExpirationDate = "05/2009",
+                },
+                ExchangeRateQuoteId = "DummyExchangeRateQuoteId-DotNet",
+            };
+
+            Result<Transaction> result = gateway.Transaction.Sale(request);
+            Assert.IsTrue(result.IsSuccess());
+            Transaction transaction = result.Target;
+
+            Assert.AreEqual(1000.00, transaction.Amount);
+            Assert.AreEqual(TransactionType.SALE, transaction.Type);
+            Assert.AreEqual(TransactionStatus.AUTHORIZED, transaction.Status);
+        }
+
+
+        [Test]
+        public void Sale_ErrorWithInvalidExchangeRateQuoteId()
+        {
+            var request = new TransactionRequest
+            {
+                Amount = SandboxValues.TransactionAmount.AUTHORIZE,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = SandboxValues.CreditCardNumber.VISA,
+                    ExpirationDate = "05/2009",
+                },
+                ExchangeRateQuoteId = new string('a', 4010),
+            };
+
+            Result<Transaction> result = gateway.Transaction.Sale(request);
+            Assert.IsFalse(result.IsSuccess());
+            Assert.AreEqual(ValidationErrorCode.EXCHANGE_RATE_QUOTE_ID_IS_TOO_LONG, result.Errors.ForObject("Transaction").OnField("ExchangeRateQuoteId")[0].Code);
+        }
+
+
+        [Test]
         public void Sale_WithRiskData()
         {
             AdvancedKountFraudSetup();
@@ -2117,6 +2161,7 @@ namespace Braintree.Tests.Integration
             Assert.IsNotNull(transaction.PayPalDetails.DebugId);
         }
 
+        #pragma warning disable 0618
         [Test]
         public void Sale_WithAllAttributes()
         {
@@ -2245,6 +2290,7 @@ namespace Braintree.Tests.Integration
             Assert.AreEqual("MEX", shippingAddress.CountryCodeAlpha3);
             Assert.AreEqual("484", shippingAddress.CountryCodeNumeric);
         }
+        #pragma warning restore 0618
 
         [Test]
         public void Sale_WithTransactionSourceAsRecurringFirst()
@@ -3279,6 +3325,8 @@ namespace Braintree.Tests.Integration
             Assert.IsNotNull(result.Target.ApplePayDetails.Payroll);
             Assert.IsNotNull(result.Target.ApplePayDetails.ProductId);
             Assert.IsNotNull(result.Target.ApplePayDetails.Bin);
+            Assert.IsNotNull(result.Target.ApplePayDetails.CountryOfIssuance);
+            Assert.IsNotNull(result.Target.ApplePayDetails.IssuingBank);
         }
 
         [Test]
@@ -3342,6 +3390,8 @@ namespace Braintree.Tests.Integration
             Assert.IsNotNull(androidPayDetails.Commercial);
             Assert.IsNotNull(androidPayDetails.Payroll);
             Assert.IsNotNull(androidPayDetails.ProductId);
+            Assert.IsNotNull(androidPayDetails.CountryOfIssuance);
+            Assert.IsNotNull(androidPayDetails.IssuingBank);
             Assert.IsFalse(androidPayDetails.IsNetworkTokenized);
         }
 
@@ -3374,6 +3424,15 @@ namespace Braintree.Tests.Integration
             Assert.IsNotNull(androidPayDetails.ExpirationMonth);
             Assert.IsNotNull(androidPayDetails.ExpirationYear);
             Assert.IsNotNull(androidPayDetails.GoogleTransactionId);
+            Assert.IsNotNull(androidPayDetails.Prepaid);
+            Assert.IsNotNull(androidPayDetails.Healthcare);
+            Assert.IsNotNull(androidPayDetails.Debit);
+            Assert.IsNotNull(androidPayDetails.DurbinRegulated);
+            Assert.IsNotNull(androidPayDetails.Commercial);
+            Assert.IsNotNull(androidPayDetails.Payroll);
+            Assert.IsNotNull(androidPayDetails.ProductId);
+            Assert.IsNotNull(androidPayDetails.CountryOfIssuance);
+            Assert.IsNotNull(androidPayDetails.IssuingBank);
             Assert.IsFalse(androidPayDetails.IsNetworkTokenized);
         }
 
@@ -5057,7 +5116,7 @@ namespace Braintree.Tests.Integration
         }
 
         [Test]
-        public void Sale_WithJCBDoesNotReturnNetworkTransactionIdentifier()
+        public void Sale_WithJCBReturnsNetworkTransactionIdentifier()
         {
             var request = new TransactionRequest
             {
@@ -5073,7 +5132,7 @@ namespace Braintree.Tests.Integration
             Assert.IsTrue(result.IsSuccess());
             Transaction transaction = result.Target;
 
-            Assert.IsNull(transaction.NetworkTransactionId);
+            Assert.IsNotNull(transaction.NetworkTransactionId);
         }
 
         [Test]
@@ -9935,6 +9994,59 @@ namespace Braintree.Tests.Integration
         }
 
         [Test]
+        public void Sale_CreateErrorTaxAmountIsRequiredForAibSwedish()
+        {
+            var request = new TransactionRequest
+            {
+                Amount = SandboxValues.TransactionAmount.AUTHORIZE,
+                MerchantAccountId = MerchantAccountIDs.AIB_SWE_MA,
+                CurrencyIsoCode = "SEK",
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = SandboxValues.CreditCardNumber.VISA,
+                    ExpirationDate = "05/2030",
+                }
+            };
+
+            Result<Transaction> result = gateway.Transaction.Sale(request);
+            Assert.IsFalse(result.IsSuccess());
+            Assert.AreEqual(
+               ValidationErrorCode.TRANSACTION_TAX_AMOUNT_IS_REQUIRED_FOR_AIB_SWEDISH,
+               result.Errors.DeepAll()[0].Code
+           );
+        }
+
+        [Test]
+        public void Sale_WithPaymentMethodTokenAndCreateErrorTaxAmountIsRequiredForAibSwedish()
+        {
+            Customer customer = gateway.Customer.Create(new CustomerRequest()).Target;
+            CreditCardRequest creditCardRequest = new CreditCardRequest
+            {
+                CustomerId = customer.Id,
+                CVV = "123",
+                Number = "5105105105105100",
+                ExpirationDate = "05/2030"
+            };
+
+            CreditCard creditCard = gateway.CreditCard.Create(creditCardRequest).Target;
+
+            TransactionRequest request = new TransactionRequest
+            {
+                Amount = SandboxValues.TransactionAmount.AUTHORIZE,
+                MerchantAccountId = MerchantAccountIDs.AIB_SWE_MA,
+                PaymentMethodToken = creditCard.Token,
+                CurrencyIsoCode = "SEK"
+            };
+
+            Result<Transaction> result = gateway.Transaction.Sale(request);
+            Assert.IsFalse(result.IsSuccess());
+            Assert.AreEqual(
+               ValidationErrorCode.TRANSACTION_TAX_AMOUNT_IS_REQUIRED_FOR_AIB_SWEDISH,
+               result.Errors.DeepAll()[0].Code
+           );
+        }
+
+        [Test]
         public void Sale_WithInvalidMerchantCurrency()
         {
             var request = new TransactionRequest
@@ -10104,6 +10216,44 @@ namespace Braintree.Tests.Integration
                 Assert.AreEqual(-5.00M, refund.RefundedInstallments[i].Adjustments[0].Amount);
                 Assert.AreEqual(Kind.REFUND, refund.RefundedInstallments[i].Adjustments[0].Kind);
             }
+        }
+
+        [Test]
+        public void Successful_ManualKeyEntryTransaction() {
+            var request = new TransactionRequest {
+                Amount = 100,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    PaymentReaderCardDetails = new PaymentReaderCardDetailsRequest
+                    {
+                        EncryptedCardData = "8F34DFB312DC79C24FD5320622F3E11682D79E6B0C0FD881",
+                        KeySerialNumber = "FFFFFF02000572A00005",
+                    }
+                },
+            };
+
+            var transactionResult = gateway.Transaction.Sale(request);
+            Assert.IsTrue(transactionResult.IsSuccess());
+        }
+
+        [Test]
+        public void Unsucccessful_ManualKeyEntryTransaction() {
+            var request = new TransactionRequest {
+                Amount = 100,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    PaymentReaderCardDetails = new PaymentReaderCardDetailsRequest
+                    {
+                        EncryptedCardData = "invalid",
+                        KeySerialNumber = "invalid",
+                    }
+                },
+            };
+
+            var transactionResult = gateway.Transaction.Sale(request);
+            Assert.IsFalse(transactionResult.IsSuccess());
+            var transactionError = transactionResult.Errors.ForObject("Transaction").OnField("merchantAccountId")[0];
+            Assert.AreEqual(ValidationErrorCode.TRANSACTION_PAYMENT_INSTRUMENT_NOT_SUPPORTED_BY_MERCHANT_ACCOUNT, transactionError.Code);
         }
 
         [Test]