+++ CHANGELOG.md
@@ -1,3 +1,42 @@
+# Changelog
+
+## 5.11.0
+- Add `PaymentMethodCustomerDataUpdated` webhook notification support
+
+## 5.10.0
+- Add plan create/update/find API endpoint
+- Add `TransactionReview` webhook notification support
+
+## 5.9.0
+- Add `LocalPaymentExpired` and `LocalPaymentFunded` webhook notification support
+
+## 5.8.0
+- Add error code `TRANSACTION_TAX_AMOUNT_IS_REQUIRED_FOR_AIB_SWEDISH` for attribute `tax_amount` in `transaction` key for AIB:Domestic Transactions in Sweden
+- Add `ExchangeRateQuoteId` to `TransactionRequest`
+- Add `EXCHANGE_RATE_QUOTE_ID_IS_TOO_LONG` to `ValidationErrorCode`
+- Add the following fields to `AndroidPayCard` and `ApplePayCard`:
+  - `Commercial`
+  - `Debit`
+  - `DurbinRegulated`
+  - `Healthcare`
+  - `Payroll`
+  - `Prepaid`
+  - `ProductId`
+  - `CountryOfIssuance`
+  - `IssuingBank`
+- Add the following fields to `PaypalDetails`:
+  - `TaxId`
+  - `TaxIdType`
+
+## 5.7.0
+- Add `PaymentReaderCardDetails` to `TransactionCreditCardRequest`
+- Add `SkipAdvancedFraudChecking` to:
+  - `Customer.create` and `Customer.update`
+  - `CreditCard.create` and `CreditCard.update`
+  - `PaymentMethod.create` and `PaymentMethod.update`
+- Add `DataOnlyRequested` to `ThreeDSecureLookupRequest`
+- Add `ChargebackProtectionLevel` to `DisputeSearch`
+
 ## 5.6.0
 - Add `PayPalMessages` to `Dispute`
 - Add `TaxIdentifiers` to `CustomerRequest`